requirements.txt
asttokens==2.0.5
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
Bcrypt-Flask==1.0.2
blinker==1.4
cffi==1.15.0
click==8.0.4
decorator==5.1.1
executing==0.8.3
Flask==2.0.3
Flask-DebugToolbar==0.11.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
ipython==8.2.0
itsdangerous==2.1.1
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
six==1.16.0
SQLAlchemy==1.4.32
stack-data==0.2.0
traitlets==5.1.1
wcwidth==0.2.5
Werkzeug==2.0.3
WTForms==3.0.1
